% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deform.R
\name{deform}
\alias{deform}
\title{Deform basemap.}
\usage{
deform(x, flatten = 0.8, rescale = c(1.3, 2))
}
\arguments{
\item{x}{an sf object (polygon or multipolygons).}

\item{flatten}{numerical value indicating the scaling in Y (1 = no flattening).}

\item{rescale}{a vector of two numeric values indicating the deformation of the top and bottom of the figure (c(1,1) = no deformation).}
}
\value{
an sf object (polygon or multipolygons).
}
\description{
Deform basemap.
}
\examples{
library(sf)
us <- st_read(system.file("us.gpkg", package="mapextrud"))
f <- getframe(us)
us2 <- deform(us)
# Example 1 (default values)
f2 <- deform(f)
plot(st_geometry(f2))
plot(st_geometry(us2), add = TRUE)
# Exemple 2 (user values)
us3 <- deform(us, flatten=0.5, rescale = c(0.5, 3))
f3 <- deform(f, flatten=0.5, rescale = c(0.5, 3))
plot(st_geometry(f3))
plot(st_geometry(us3), add = TRUE)
}
