% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrude.R
\name{extrude}
\alias{extrude}
\title{Extruded map (beta version)}
\usage{
extrude(
  x,
  var,
  k = 1,
  lwd = 1,
  col = "white",
  border = "black",
  regular = FALSE,
  add = FALSE
)
}
\arguments{
\item{x}{an sf object (polygon or multipolygons).}

\item{var}{name of the numeric field in df to plot.}

\item{k}{numeric value ti adjust the height of the extruded polygons}

\item{lwd}{line thickness}

\item{col}{One color or the field name containg colors for each polygons}

\item{border}{color border}

\item{regular}{use TRUE with regular grids for faster calculation}

\item{add}{whether to add the layer to an existing plot (TRUE) or not (FALSE).}
}
\value{

}
\description{
Extruded map (beta version)
}
\examples{
library(sf)
us <- st_read(system.file("us.gpkg", package="mapextrud"))
basemap <- deform(us)
extrude(basemap, var = "pop2019" , k = 1, col = "white")
}
