% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getframe.R
\name{getframe}
\alias{getframe}
\title{Get Frame.}
\usage{
getframe(x)
}
\arguments{
\item{x}{an sf object.}
}
\value{
an sf object.
}
\description{
Make a frame around the map.
}
\examples{
library(sf)
us <- st_read(system.file("us.gpkg", package="mapextrud"))
f <- getframe(us)
plot(st_geometry(f))
plot(st_geometry(us), add = TRUE)
}
