% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legendextrud.R
\name{legendextrud}
\alias{legendextrud}
\title{Legend}
\usage{
legendextrud(
  x,
  var,
  k,
  title.txt = "Title of the legend",
  unit = "",
  title.cex = 0.8,
  values.cex = 0.6,
  pos = NULL,
  title.font = 2
)
}
\arguments{
\item{x}{an sf object (polygon or multipolygons).}

\item{var}{name of the numeric field in df to plot.}

\item{k}{numeric value to adjust the height of the extruded polygons}

\item{title.txt}{Title of the legend}

\item{unit}{unit of the variable}

\item{title.cex}{title cex}

\item{values.cex}{velues cex}

\item{pos}{c(x,y) position on the legend}

\item{title.font}{title font}
}
\value{

}
\description{
Legend
}
\examples{
library(sf)
us <- st_read(system.file("us.gpkg", package="mapextrud"))
basemap <- deform(us)
extrude(basemap, var = "pop2019" , k = 1, col = "white")
legendextrud(x = basemap, var = "pop2019", k = 1, title.txt = "Population in 2019", unit = "inh.")
}
