% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate.R
\name{rotate}
\alias{rotate}
\title{Rotate basemap.}
\usage{
rotate(x, deg)
}
\arguments{
\item{x}{an sf object.}

\item{deg}{degree angle.}
}
\value{
an sf object.
}
\description{
Rotate sf basemap.
}
\examples{
library(sf)
us <- st_read(system.file("us.gpkg", package="mapextrud"))
us2 <-rotate(us, 180)
plot(st_geometry(us2))
}
